openkms-backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.30.0
alembic==1.12.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2

# Caching
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development Tools
python-dotenv==1.1.1
jinja2==3.1.6

# Optional - for production
gunicorn==21.2.0
uvicorn[standard]==0.24.0